String Decryption, Ida database
Adding a tool for decrypting strings however I'm not familiar with ida python or python in general. so you have to enter strings manually.

In the database there is a func named Main2, it looks like some forum of memory communication by writing to something however it is not used